Install add-on files first and invert the uninstallation
Extensions bundled in add-ons might require the bundled files so those
should be installed first (for example, add-ons with script types and
templates, the templates need to be present when the extension is loaded
otherwise the templates are not read when the script types are
registered).
The uninstallation should be the inverse of installation to prevent
removing anything that might be in use by a previous unloaded component.