Ignore add-ons that are not zip files or are corrupt
When an add-on file is found it's read its ZapAddOn.xml file which might
lead to exceptions when the file is corrupt (or simply not an add-on).
When that occurred the add-on would still be considered for further
processing leading to further exceptions (for example, when searching
for classes during initialisation) and it would be shown as an installed
add-on (showing just its ID).
Possible exceptions, when reading the ZapAddOn.xml file, were being
caught and caller code were not notified or aware that the file was not
a valid add-on. The change lets the exception propagate up in call chain
so callers can ignore those files.