Hide splash screen and inform user about starting errors
If an Error or (uncaught) Exception is thrown during initialisation the
main thread, responsible for the intialisation, dies but the splash
screen is still displayed, until the user manually closes it.
This change closes the splash screen when that happens and, when in non
dev build, informs the user about the error (with the details for
reporting).