Prevent spider's memory leak when in the presence of custom parsers
The SpiderController adds itself, as a listener of resources found, to
the parsers set, though it never removes itself after the spider process
has ended. With default parsers there are no problem since they are
instantiated by SpiderController and are not referenced by other
classes. With the custom parsers, set by external classes, the lifetime
is undefined those parsers might be referenced by classes whose lifetime
is greater than the spider process which leads to a memory leak. This
problem could be reproduced with the custom parser WSDLSpider added by
"SOAP Scanner" add-on.