Issue 1502 - Allow to remove script engine wrappers
Only remove default templates when removing a script engine wrapper. Fix
NullPointerException and ClassCastException when un/setting a script
engine. Disable the scripts when engine is removed and restore its state
when the engine is re-set. Always set the name of the engine of a script
wrapper, to know what engine it has when its engine is set to null.