Issue 1515 - Not all "Zest" components are unloaded during uninstallation
Update ScriptUI interface to allow to remove all added/set components,
handle TransferHandlers and be notified of engine registration/removal.
Expose the ExtensionScript methods isSameScriptEngine and
hasSameScriptEngine as those are useful for other extensions/add-ons.
The notification of registration/removal of an engine is commented until
the known implementation of ScriptUI implements those methods.