Issue 1208 - Search classes/resources in add-ons declared as dependencies
Add-ons can depend on each other (no cyclic dependencies allowed) and
depend on a specific Java version. The AddOnLoader will not run add-ons
whose dependencies are not met (but will be run once are met, for
example after installation of a dependency). The update of dependencies
require a temporary uninstallation of the dependent add-ons, to properly
link the new classes, used by the dependent add-ons.
The AddOnClassLoader searches for classes in the dependencies before
failing with a ClassNotFoundException, similarly for resources.
The check for updates will automatically download any required
dependencies to fulfil the add-ons being updated, unless there's a
dependency issue (for example, an add-on is no longer runnable or the
update uses a newer Java version).
The dialogue "Manage Add-ons" reports any dependency issues in the
tables, by showing a warning icon, with tool tip containing the details
of the issues. Both add-ons not installed as updates might show running
issues. Any required change, not explicitly selected by the user, will
be reported and done only after confirmation (for example, inform about
install of dependencies of a selected add-on). The user will be informed
of unsaved resources and active actions before uninstalling/updating an
add-on.
Extension can have a semantic version, reference the corresponding
add-on and report active actions (e.g. scans).
During the removal of an extension the menus and menu bar are
revalidated, to prevent removed items from being shown.
After ZAP start, if a previously running add-on is no longer runnable,
(for example, because of a missing dependency) it will be reported (with
a information dialogue, or a log message when view is not initialised).
The uninstallation of the add-ons is no longer done in the EDT, instead
is shown a progress dialogue. Also it's caught Throwable instead of
Exception to prevent the incorrect propagation/notification of the
failed uninstallation.
The active actions of the extensions are reported, as the unsaved
resources, when shutting down ZAP.