Issue 1530 - Default attack strength and alert threshold, set in 'Scan Policy', not used by scanners
Fix the AbstractPlugin.cloneInto by getting the actual values of attack
strength and alert threshold instead of final value (when in DEFAULT it
would return, always, MEDIUM, the default value, instead of DEFAULT).
Set the default attack strength and alert threshold to the scanners
(Plugin) of the ScanPolicy when those properties are changed to keep the
defaults of the scanners up-to-date and remove the defaults set in
HostProcess since are no longer needed (and had no effect, since the
values previously set by the clone were never DEFAULT).