Do not log as error if scanners have same ID/name but different quality
When loading scanners it might happen, by mistake, that different
scanners have the same ID, in those cases is logged an error message
indicating that problem. Although in some cases the same scanner but
with different quality status might be loaded at same time (e.g. when a
scanner is promoted to higher quality status, alpha to beta or beta to
release, and just one of the add-ons containing the scanner is updated
or if one add-on is updated only after updating the other, instead of
both at same time) in those cases it makes more sense to not log an
error since it's (apparently) the same scanner, moreover it should be
checked and loaded the scanner with higher quality status (instead of
relying in the order they are added).
The changes done to PluginFactory now check if the scanner with same ID
has the same name and in those cases check also the quality status and
use the one with higher quality. It's also logged a message with info
level containing the details (i.e. id, canonical class name and quality)
of the scanners.