Do not run active scanners that do not target the enabled technologies
Add a method to Plugin interface, boolean targets(TechSet), to allow to
query the scanners to know if they target the enabled technologies, if
not they are not run.
Change AbstractPlugin to implement the new method.
Change HostProcess to call the new method and mark the scanners as
skipped if they do not target the enabled technologies.
Update the JavaDoc of related methods in Plugin and AbstractPlugin.
Core changes for #1618 - Target Technology Not Honored