Add a table with common behaviour for use in, mainly, status panels
The table allows to choose to auto-scroll on new values and selects the
entries on mouse context menu invocation.
The auto-scroll can be enabled/disabled in the table's options (context
menu invoked with the button above scroll bar).
Change HistoryReferencesTable to extend from the new table and remove
common code.