ascan: make sure technologies of the context are used, if not overridden
Change CustomScanDialog to set the selected context to the target (if
not already set) to ensure that it's used later on (for technologies as
well other functionalities like scan in context's scope).
Change ActiveScanController to set the technologies of the context to
the active scan, if not overridden.
Fix #1735 - Context's technologies not used in active scan unless
overridden