Change context's action "newContext" to require parameter contextName
The action no longer does anything if the context name is not given
(because of changes to fix an issue in renaming of contexts). Changing
the parameter to be mandatory reflects better the new behaviour.
Update documentation of the action to say that creates the context with
the given name.
The change also fixes an exception that occurred when the parameter was
not present, 'JSONException: JSONObject["contextName"] not found.', as
the parameter's value was being accessed before checking if it was
present. The "if" statement that checked that the parameter was not
empty to create the context was removed, that's already enforced by core
implementation for mandatory parameters.