Remove redundant and duplicated code in class ContextAPI
Remove redundant checks for existence of parameters (methods
getRegex(JSONObject) and getContextName(JSONObject)), the parameters are
already mandatory thus guaranteed to exist.
Remove duplicated code, in action "exportContext", by using a function
already available, that returns a context with a given name and replace
the code in getContext(String) that searched for a context
with a name with Session.getContext(String), which does exactly the same
thing.
Change to search for the context only once in the action
"setContextInScope" and remove redundant check for existence of
parameter (it's mandatory).
Add JavaDoc to getContext(JSONObject) which explain the preconditions
(the parameter contextName must exist, which it does if the parameter is
mandatory).