Fix the report of newer versions after check for updates on start
Change ExtensionAutoUpdate.gotLatestData(AddOnCollection) to initialise
the "manage add-ons" dialogue so that it gets notifications of possible
ZAP/add-on changes and is also shown when needed and change
checkForAddOnUpdates(AddOnCollection, OptionsParamCheckForUpdates) to
not keep checking for more changes if the dialogue is shown when there's
conflicts in update of add-ons (e.g. requires uninstallations or newer
Java version) and to show the dialogue if no updates were installed when
check for add-on updates is enabled.
Fix #1763 - Automatic check for updates fails to report new versions