Use ILLEGAL_PARAMETER to indicate issues in values of API parameters
Change the ApiException.Type used in exceptions thrown by APIs from
BAD_FORMAT to ILLEGAL_PARAMETER when indicating issues in values of
parameters.
The BAD_FORMAT is used to indicate that the requested format for the
response is invalid (i.e. not JSON, HTML, XML...).
In the some cases an Exception was being caught to report issues in the
parameters, although it was also catching internal errors, in those
cases the catch block was split into two to differentiate internal error
(INTERNAL_ERROR) from parameter issues.
Document BAD_FORMAT and ILLEGAL_PARAMETER.