Deprecate FileXML and change Session to use ZapXmlConfiguration
Deprecate FileXML in favour of using ZapXmlConfiguration as it provides
more features and it's more maintained.
Change Session to use ZapXmlConfiguration and no longer extend from
FileXML. The change breaks binary compatibility although it's very
unlikely to affect 3rd party code given the nature of the Session class
(core) and the public methods of FileXML (methods to get/set the
Document and read/save from/to a file; read and save was already being
done internally by Session class).