Keep AbstractPlugin's enabled state in sync with AlertThreshold state
A plugin might be disabled be setting its AlertThreshold to OFF, though
the enabled state was not being updated accordingly, so client code that
queried the enabled state directly (AbstractPlugin.isEnabled()) would
see a different result than checking if the AlertThreshold was OFF.
This would lead to wrong results when checking if the dependencies of a
plugin were enabled or not, appearing that the dependencies were enabled
when that was not the case (AlertThreshold was OFF but enabled state was
true).
The change done to AbstractPlugin ensures that the enabled state is
always in sync with AlertThreshold state, preventing the above
inconsistencies.
The save/load of the plugin's enabled state to a file was changed to be
always performed, the check done previously would prevent the save/load
since the config used for the state was never null, it is created if not
present in the config file. Not saving/loading the state would also
contribute for inconsistencies (if the enabled state was not kept in
sync).
For a matter of consistency the code was changed to always use the
method isEnabled() to check if the plugin is enabled or not.
Remove, now redundant, methods in ActiveScanAPI that changed both the
enabled and AlertThreshold states (to keep them in sync).
Fix #1785 - Plugin enabled even if dependencies are not, "hangs" active
scan