Change ScanPanel2 to sort the scan entries using the ID of the scan
Replace the use of String with a custom class that does the comparisons
using the ID of the scan, it also keeps the scan (GS) for easier
retrieval when obtaining the selected scan. Remove code that's no longer
need by keeping the scan in the custom class.
Add a new instance variable to keep the "Select Scan" entry, it's used
very often.
In the method scannerStarted(GS) a check was removed because it was
redundant, the scan entry was always present in the model.
Fix #1788 - Scan Progress Pane Needs Sorting Change