Correct a check that prevents the alerts from being added to the report
Split the "if" statement in ExtensionAlert.addAlertToTree that checked
if ZAP is in "low memory" or in daemon mode. Both states are unrelated
and, when in daemon mode, prevented the alerts from being added to the
corresponding SiteNode (which is from where the alerts are obtained for
inclusion in the report).
Fix #1792 - Report not generated in daemon mode