Change API to return error response in requested format
Change API to catch Exception instead of ApiException so it can also
return a response in those cases (instead of closing the connection
abruptly).
Change to check what kind of exception and type in case of ApiException
is begin caught to log the error with a more appropriate level (error vs
warn). In case of ApiException with type INTERNAL_ERROR it's not logged
anything, as those exceptions were already handled by ApiImplementors
which should have logged the exception/issue.
When the requested format is OTHER it's not generated any response, the
exceptions should be explicitly handled by the ApiImplementor because
the format is not known (so no suitable response can be generated).
Prepare the API to return different status code for client/server
errors.
Default to HTML format if the format is not known (BAD_FORMAT).
Fix #1805 - ZAP API might not return the response in requested format on
errors