Fix equals check in API that evaluated always to false (different types)
Change ApiException to expose its Type so that it can be correctly
checked in API class when the exceptions are handled.
The issue would lead to unnecessary log of exceptions of type
INTERNAL_ERROR (which are supposed to be already logged).
Rework the constructors of the ApiException to call other constructors
with null values, to reduce code duplication.