Increase error count just once per logged error
Move the increase of the error count to after the EDT check.
If the logging does not occur in the EDT the method is executed again
but in the EDT, the problem was that the error count was increased
before checking if the current thread was the EDT, leading to count of
the same error twice.