spider: restore acceptance of selected context in Spider dialogue
Change SpiderDialog to not validate (nor try to use) the text, in the
input box, as URL if it corresponds to the text representation of a
selected Target (for example, a context or a regular node of the tree).
Change StandardFieldsDialog to expose a method, getTargetText(Target),
that creates the text representation of a Target to allow the previous
checks.
Simplify if/else statements in SpiderDialog.save() by using a List,
instead of an array, to keep the context's specific objects (which are
now added as needed).