spider: change Spider tab to allow to auto-scroll on new entries
Change class SpiderPanel to use ZapTable (instead of JXTable) as it
supports auto-scrolling on new values (it's enabled by default).
Remove now redundant call that sets the column control visible.
Replace the use of a custom MouseAdapter to show the context menu with
Swing built-in support of context menus (allows to be invoked with mouse
and keyboard, where supported, moreover ZapTable's implementation makes
sure that the entries are automatically selected on context menu
invocation with mouse).
Fix #1593 - Auto-scroll in Spider tab