spider: synchronise calculation and notification of the progress
Change Spider class to synchronise the method postTaskExecution() where
the progress is calculated and interested parties notified, otherwise it
has a concurrency issue that might lead to notification of wrong
progress, for example:
 - Thread A calculates the progress as 97;
 - Thread B calculates the progress as 100 (spider process is finished);
 - Thread B notifies (synced method notifyListenersSpiderProgress(...));
 - Thread A notifies;
 - Listeners would have last progress as 97, instead of expected 100.

Change the type of the instance variables "tasksTotalCount" and
"tasksDoneCount" to int (from AtomicInteger), all accesses are now
synchronised.
Correct JavaDoc of postTaskExecution() to indicate that it's called
after the execution of the task.
Fix #1858 - Spider might report wrong progress after finishing