Move creation/loading of ZAP icon images to a utility class
Change classes AbstractDialog, AbstractFrame and SplashScreen to use the
new method DisplayUtils.getZapIconImages(). The images were/are already
being cached in memory by Toolkit implementation so this change just
removes the code duplication (and processing) of creation/loading of the
images.
Deprecate AbstractFrame.loadIconImages() in favour of using the method
DisplayUtils.getZapIconImages().