Fix memory leaks in ExtensionHistory and ExtensionAntiCSRF
Change ExtensionHistory to use a ReferenceMap (value as SoftReference),
instead of a Map, to cache the HistoryReference(s), so that they can be
reclaimed when the memory is really needed, also do not cache temporary
active scan requests, those do not need to be cached (not accessed by
other modules of the application).
Change ExtensionAntiCSRF to persist the HTTP messages of the
AntiCSRFToken(s) so that they do not lead to exhaustion of memory (the
messages are read (and cached) from database as needed).
Removed unused instance variable, urlToToken, and related code from
class ExtensionAntiCSRF.
Fix #1890 - ZAP can't completely scan OWASP Benchmark