Make sure AbstractPlugin enabled state is kept the same when cloning
Switch the order in which the alert threshold and enabled state is set
when cloning so that the former does not change the enabled state (e.g.
in case it's non OFF and the plugin is disabled, which would enable it).
Add tests to assert the expected behaviour.
Fix #1893 - Can't disable and enable specific scanners through the API