Log that the root CA certificate was created
The creation of root CA certificate, if successful, just logs "Creating
new root CA", which is very misleading giving the impression that ZAP
hanged, as reported in #1920.
With the change it always logs something either that it was created or,
if something went wrong, the exception.
Tweak existing log messages to include "certificate", also change log of
the exception to explicitly say that the creation of root CA certificate
failed.