Add ZAP CLI to Docker images
Add ZAP CLI to Docker images. ZAP CLI can be used to run quick, targeted
scans with ZAP from within the container. It can be used with the API key
disabled or a custom API key, so that it can be used with the current
version of ZAP in the container.