Do not attempt to add a script with missing type, when loading ZAP
Change ExtensionScript#postInit() to only (attempt to) add the scripts,
loaded from the configuration file, if the type is not missing and log,
with WARN level, otherwise.
Change to also catch the exception thrown if unable to add the script,
in case the script configuration is not correct (for example, user tries
to add a script with duplicated name through the command line args).
The user will also be informed, through a dialogue, which of the scripts
were not loaded.
Fix #1949 - Script with missing type prevents scripts and templates from
being loaded