Inform if the authentication script was not yet loaded when adding users
Change AuthenticationMethod to allow to validate the creation of users
(i.e. AuthenticationCredentials).
Change ContextUsersPanel to validate the creation of users before
showing the dialogue to create them.
Change ScriptBasedAuthenticationMethodType to, when validating, inform
and prevent the creation of users if the script was not yet loaded.
Tweak resource messages to differentiate from script not yet loaded and
a script not yet loaded nor configured.
Fix #1951 - Exception when trying to add users before loading an
authentication script