Load plugins' state when loading all plugins from a policy file
Change PluginFactory.loadAllPlugin(Configuration) to call the method
Plugin.loadFrom(Configuration) to make sure that the plugin's state is
loaded from the policy file.
Change AbstractPlugin.saveTo(Configuration) and loadFrom(Configuration)
to save/load the in-memory enabled state, otherwise the AbstractPlugin
would not save the (current) enabled state and/or would be kept enabled
even if it was disabled in the provided policy file.
Add tests to AbstractPluginUnitTest to assert the expected behaviour.
Fix #1971 - Active scanners' enabled state not correctly restored from
policy file