Change CustomScanDialog to not cache ScanPolicy, once closed
Change CustomScanDialog to stop referencing the ScanPolicy created for
the dialogue once the dialogue is closed (or hidden).
The dialogue now just caches the name of the selected policy, obtaining
the ScanPolicy when the dialogue is (re)initialised.
Change PolicyManager class to expose the method that allows to check if
a policy exists (policyExists(String)), to handle the cases when the
policy is removed/renamed.
Fix #1960 - Active Scan dialogue might use outdated scan policy