Notify listeners that the script was changed when changing enabled state
Change ExtensionScript.setEnabled(ScriptWrapper) to notify the script
listeners (ScriptEventListener) that the script was changed.
The script (ScriptWrapper) is marked as changed when it's enabled or
disabled, but the script listeners were not notified of the change, for
example, causing the UI to not be immediately updated to reflect the
change by enabling the save button.
Fix #1939 - Scripts - Save Button Enablement Issue