Allow to spider a context through the ZAP API
Change the action "scan" of SpiderAPI to allow to specify the name of
the context used in the spider process.
Add methods to ApiUtils that allows to obtain the context from the API
parameters and by just using the context name (which throws an
ApiException if it doesn't exist).
Change ContextAPI to use the new method from ApiUtils.
Fix #1953 - Allow to spider a context through the ZAP API