Use table model row when showing the modify dialogue on double click
Change AbstractMultipleOptionsBaseTablePanel to convert the table view
row index to model row index before showing the modify dialogue.
The view and model table row indices might differ when the columns of
the table are sorted, which could lead to a different entry being shown
than the one double clicked.
Move the if statement that guards against negative rows to MouseAdapter,
the check is now done (and required) before converting the row indices.