Require all files downloaded from the marketplace to have a hash
Change class Downloader to not download the file if it does not have a
hash (and of the expected type "<ALGORITHM>:<HASH>").
Change to not validate the hash if the download was cancelled, the
validation of the hash would fail, moreover the file is deleted.
Extract two methods, one for the download of the file and another for
the validation of the hash of the downloaded file.
Initially the marketplace didn't provide the hash of the available files
so, to allow an easier transition older files without hash were still
accepted, while newer files with hash were validated and discarded
if the hash didn't match. Now that all files in the marketplace have a
hash, the exception done for older files is no longer required.