Install dependencies of the add-on installed through the command line
Change ExtensionAutoUpdate to install the dependencies of the add-on
being installed/updated through the command line, aborting the
installation if it's required to uninstall add-ons (for example, because
of incompatibilities in the dependencies of the add-on(s)).
Inform about all the add-ons being installed (downloaded) and possibly
updated (because of dependency requirements). Also inform, as error,
about failed downloads.
Add helper method AddOnDependencyChecker.calculateUpdateChanges(AddOn),
not requiring the callers to create a Set when just calculating the
update changes of a single add-on.