Fix issues with implementation of Hirschberg's algorithm
Correct the implementation of HirshbergMatcher.algC(...), by appending
the character when the characters of the strings are the same in step
"m == 1". It was being created a StringBuilder with the character as
parameter which was being used as int thus, instead of "appending", it
was setting the initial capacity of the StringBuilder. Because of that
the method HirshbergMatcher.getLCS(...) was returning always an empty
LCS.
Fix StackOverflowError exception that occurred when obtaining a LCS if
string A was empty (change to return, immediately, an empty string if
that's the case).
Add tests to assert the expected behaviour.