Save the Context after configuring the authentication through the APIs
Change AuthenticationAPI, ForcedUserAPI, SessionManagementAPI and
UsersAPI to call the method Context.save() after doing the changes
to save the authentication configurations of the context.
Change UsernamePasswordAuthenticationCredentials to cope with null
username and password when persisting the data (happens when no valid
authentication credentials were yet set to a user).
Fix #2052 - Authentication changes done through the API not saved to
session