Fix ClassCastException in ClassLoaderUtil
Change ClassLoaderUtil to not attempt to cast/use the default delegation
parent class loader as URLClassLoader if it's not one, otherwise it
would lead to a ClassCastException and ZAP would fail to start. ZAP does
not require the dynamic loading of the dependencies or other resources,
if they are already in the classpath (or set in the ClassLoader, which
is the case when the ClassCastException happened).