Allow add-ons to specify which classes can be loaded by its classloader
The change allows the add-ons to specify (through its ZapAddOn.xml file)
which classnames are allowed or restricted for the whole add-on and for
its extensions, when loading classes from the add-on.
The change allows add-ons to better compartmentalise its extensions,
for when they depend on other add-ons, preventing the load of the same
classes (from the add-on itself) more than once (which leads to
misleading issues since the classes are the same, but have different
class loaders).