Bump version and regenerate Python API client to include latest API code
Regenerate to include the options of the AJAX Spider add-on, now exposed
through the API.
Add the reveal API to ZAPv2 class (already included but not directly
accessible through the main class).
Bump version to 0.0.6 (latest release is 0.0.5, though the code didn't
reflect that) and update the download_url to the latest ZAP version
(2.4.3).