Fix NullPointerException on missing context's authentication script
Change the method ScriptBasedAuthenticationMethodType.loadMethod
(ScriptBasedAuthenticationMethod, List<String>, List<String>) to not
continue loading the script if it was not found, preventing the
NullPointerException, also, if in GUI inform about the missing script.
Change the method ScriptBasedAuthenticationMethod.duplicate() to only
copy its parameters' values if those are non null (the parameters'
values are null, for example, if the script was not found or immediately
after instantiation).