Prevent active scan from becoming in undefined state on exceptions
Change method HostProcess.run() to always notify the listeners that the
process has finished. Also change to catch Exception, log it as an
active scanning error and stop the process (to prevent scanners from
keeping running).
Change method HostProcess.processPlugin(Plugin) to always notify the
listeners that the plugin has finished, when traversing the tree
(otherwise, in case of an exception, the scanner would be reported as
running even if not).
Continuation of #2124 - Prevent active scans from becoming in undefined
state on exceptions