add-marketplace-sorting
Setup auto row sorter for installed and uninstalled addons tables in the
marketplace dialog. This is most handy in the uninstalled section if you
want to change to name sorting (vs. the default quality-name), or if (in
either table) you want the list reversed.

Add StatusUI component to be leveraged elsewhere for status/quality 
info. View now initializes a map (AddOn.Status, StatusUI), and exposes 
a getStatusUI method.