Change AbstractParam's clone method to clone its configuration
Change to clone the configuration instead of copying it because the
latter does not handle hierarchical configurations (the ones used by
ZAP) correctly, leading to missing options in the new configuration.
This was making the AJAX Spider lose click elements when the options
were being cloned.
Add tests to assert the expected behaviour.
Fix #2151 - AJAX Spider does not click all elements set in the options