Expose the HttpSender to AuthenticationScript implementations
Change class AuthenticationHelper to expose the HttpSender to the
AuthenticationScript implementations, so that the scripts can configure
how the messages are sent/handled.
Update template script, "Authentication default template.js", with some
examples on how to configure the HttpSender.
Fix #2070 - It should be possible for the authentication scripts to
configure how the messages are sent