Allow authentication scripts to provide logged in/out indicators
Change AbstractAuthenticationMethodOptionsPanel to allow authentication
implementations to control the logged in/out indicators through an
interface, AuthenticationIndicatorsPanel, which allows to set the logged
in/out indicators, enabled/disable the fields and set its tool tip.
The interface is implemented internally by ContextAuthenticationPanel
and provided to the authentication method implementations when bond,
loaded into the view.
Add an interface, AuthenticationScriptV2, that allows the Authentication
script implementations to provide the logged in/out indicators through
getters (getLoggedInIndicator() and  getLoggedOutIndicator()).
Change ScriptBasedAuthenticationMethodType to check if the
authentication scripts implement the new interface, in which case it
loads the indicators, otherwise it does nothing (new), and keeps the old
behaviour. When bond to the view it disables the fields of the logged
in/out indicators, as those are already being provided by the script
(it also shows a tool tip indicating that).
Update the default authentication template script with example functions
on how the new interface is expected to be implemented.
Fix #2068 - LoggedIn and LoggedOut indicators inside authentication
scripts